Some refactoring. Started work on Board Api.
