+++ py3wetransfer/__init__.py
@@ -1,1 +1,2 @@
-from .base import WeTransfer
+from .board import Board as BoardApi
+from .transfer import WeTransfer as TransferApi
